more portability fixes.
made sure the slice number is always passed as an integer, instead of
sometimes int and smetimes uint8_t

do not include mmx.h when on non-X86 builds (because of the nonstandard
long long in there)

+++ trunk/mpeg2dec/libmpeg2/mpeg2_internal.h
@@ -197,7 +197,7 @@ extern mc_functions_t mc_functions_altivec;
 extern mc_functions_t mc_functions_mlib;
 
 /* slice.c */
-void slice_process (picture_t * picture, uint8_t code, uint8_t * buffer);
+void slice_process (picture_t * picture, int code, uint8_t * buffer);
 
 /* stats.c */
-void stats_header (uint8_t code, uint8_t * buffer);
+void stats_header (int code, uint8_t * buffer);
